[WIP] Registering Umb package
Attempting to manually add the package to the "installed packages"
in the Umbraco back-office.

The scenario behind this is if you install via NuGet. Then the
back-office has no idea that the package is installed.

By having it recognised as "installed", is that it could be un-installed
and the "options" page could (potentially) be set too.