DataList - removing strongly-typed
**BREAKING CHANGE**

FWIW, (by some) strongly-typing is considered the goal,
but comes at a cost of using Reflection (and often an
over-engineering effort).

Reviewing how the DataList editors and sources were being used,
I'd notice that de-serializing the JSON config value back to the
POCO class was causing it to mutate.

The `GetItems` method was relying on the POCO's properties, rather
than independently deal with the config values.
If the POCO property values were mutated between iterations,
this is very problematic!

The better course of action is to remove the POCO properties
(decorated with `ConfigurationField` attribute) and manually outline
the desired fields.

In a way this is weakly-typing the configuration, but it does reduce
the amount of de-serialization and use of Reflection.

This is quite a big breaking-change, but will mitigate against
potential future headaches.

_(and Contentment isn't officially released yet... right?)_