:warning: ConfigurationEditor - BREAKING CHANGE! :warning:
In `ConfigurationEditorModel`, I've renamed the `Type` property to `Key`
as "Type" originally made sense when it was only dealing with POCOs,
but since ContentBlocks, it's been adapted to work with other data.
"Key" makes more sense.

I say this is a BREAKING CHANGE because the model has changed,
but from a user's perspective, I have patched the relevant code
so there is no data loss.

:warning: However if you are doing anything adventurous using
ConfigurationEditor in your own custom editors, then you'll need
to tweak your code (from using `Type` to `Key`) and recompile it.
But given that Contentment has not officially been released yet,
I should be okay to make breaking-changes.