Removed Toggles list editor
Once the `umb-toggle-group` directive became available in Umbraco core,
I was eager to make a list editor from it. However since developing it,
I've never used it - as the difference in usage from a regular Checkbox
List is next to nothing.

Apologies to any early adopters who may have secretly enabled this list
editor. Feel free to take the code and use it in your own
projects/implementations.

It's not something that I want to support ongoing.