Removed Cards list editor
I'd prototyped the Cards editor when I was looking for familiar Umbraco
 UI components that I could reuse. It turns out that since I developed
it, I've never used it! I can't see when I'd use this compared to the
Item Picker list editor.

Apologies to any early adopters who may have secretly enabled this list
editor. Feel free to take the code and use it in your own
projects/implementations.

It's not something that I want to support ongoing.