DataList ValueConverter - fixed bug when ModelsBuilder is not enabled.
Fixes #56

Turns out that when ModelsBuilder is not enabled, the
`NoopPublishedModelFactory` is used. This sets the `CrlType`
to be `typeof(IPublishedElement)` (I've no idea why!), so this wasn't
compatible with the logic for DataList's ValueConverter, so strongly-
typed values were being returned as comma-separated strings.

Whilst refactoring the ValueConverter, it turns out that I didn't need
to know the property's `CrlType`, but only whether the target value
was intended to be multiple (`IEnumerable<>`) or not.