DataList ValueConverter - fixed bug
If a list-editor is configured to have a single value, but multiple
values have already been selected, e.g. it's been re-configured after
data is already set. Then the value-converter was returning a
comma-separated value, but that doesn't cast to the target
object-type correctly.

To fix this, I take the first value of the enumerable and convert
that to the target object-type.