TextInput - amended ValueConverter
I was inheriting from Umbraco's `TextStringValueConverter`,
but the recently DependencyInjection updates made a breaking-change
for me, one of those "once bitten, twice shy". I've decided to detach
from it and go standalone. Less moving parts.

The main loss is the URL parsing feature, but I'm not sure that
parsing `href` attributes for `localLink` tags is useful for plain
textstring editors.