C# coding style, explicit ` == true`
I can appreciate that this coding style could drive some developers
up the wall, (and may consider it redundant code), these are typically
the same developers who never adopted the use of `var`.

For me, it improves readability (especially when reading code outside
of Visual Studio, e.g. on GitHub directly).

In terms of readability, it's about cognitive load, e.g. use of
`TryGetValue` makes the presumption that you already understand the
method's return value, (implicit coercion), whereas with the `== true`
condition, you can see it's a boolean check.

In terms of performance, I don't buy into the "redundant code" reason,
as the C# compiler will optimize for this explicitly.

After all that, who are we writing this open-source code for?
A compiler or a human?