Telemetry - captures additional configuration
Since introducing the telemetry feature in v1.2.0, DataList appears to
be the most popular, but I had no idea which types of data-source or
list-editors were being used. This amendment attempts to capture the
minimum amount of data that I need to analyze the usage of DataList
(and ContentBlocks, TextInput) configurations.