:shushing_face:
Reverted my decision to have unlockable passcodes.

My primary reason was to have a mechanism for users to access
"exclusive" content (e.g. data-sources), but I've since come to realise
that isn't my goal for Contentment. First and foremost I want to share
code with the world.

The secondary reason to hide all these extra data-sources was to not
overwhelm new and casual users. The extra ones could be considered
advanced or edge-case, they aren't data-sources for typical use.
(Some folk may disagree with me on that point, but hey, it's my project)

My (current) happy compromise is to keep the new extra data-sources
hidden, and still unlockable with a code snippet.

e.g. `composition.UnlockContentment();`

I've also made all the C# classes for the extra data-sources to be
`public`, so if there is a specific extra data-source needed, it can be
done using...

e.g. `composition.ContentmentListItems().Add<TimeZoneDataListSource>();`

Whether any of this is a good decision or not, time will tell.
It's all progressive.