UmbracoImageCrops data-source - swapped picker to use radiobuttons
Using the ItemPicker was part of my original prototype, but after
usage, since it's the only field in the configuration, swapping it to
use a radio-button-list makes more use of the available space.

Also, in doing this, the value is swapped from being an array to
a single string value.