DataListValueConverter - workaround for BlockList issue
It appears that Umbraco Element Property (that BlockList uses) does not
make a call the `ConvertSourceToIntermediate()` method. Meaning that the
`inter` value is in the `source` format.

Most likely it needs to be reviewed in Umbraco core, but as a workaround
I've added a check for the source format, (in this case a `JArray`).

Resolves #111

(cherry picked from commit e03cb2114e8691f06af359a9b06017bc8197002a)