⚠  [BREAKING-CHANGE]
DataList value-converter - change of PropertyValueType

Following on from the refactor in commit 6bac83c157441aa889680e2a4e9ee7355ed8c914.
The excepted return object-type was an `IEnumerable<>`, however it was
actually an `Array<>`, meaning that it had always been a mismatch.
Changing this to be a `List<>` would still be compatible with
`IEnumerable<>`, but I wasn't happy with the mismatch, so wanted to make
 it consistent and return as a `List<>` object-type.