ContentBlocks - renamed private "fake" controller class name
An attempt to resolve an issue raised in #115.

The previous name `FakeController` may have been conflicting with an
Umbraco private controller class name. I'm unsure if this will resolve
the underlying issue, but there is no harm in renaming this class.

I also changed the base class to be `ControllerBase`, following the
same as Umbraco's `FakeController` does.
https://github.com/umbraco/Umbraco-CMS/blob/release-8.14.0/src/Umbraco.Web/Mvc/ControllerExtensions.cs#L162