Reworked CSS/JS assets
Previously, I'd been bundling/minifying the CSS and JS so that they
would be production-ready AND that the number of HTTP requests
in the CMS backoffice would be reduced during development.
Given that Smidge will bundle any pkg assets (when debug=false), the
overhead of multiple HTTP requests doesn't seem too detrimental for v9.

An advantage of this is that now the individual property-editor JS
files are available, it may encourage collaboration. I have a feeling
that bundling/minifying put folk off from digging into the codebase.

This change also means that I'm removing the `WebAssets` code and
deleting the AjaxMinifier.exe tool (goodbye old chum).