Moved the startup logic back to the Composer class
I've had a dilemma on how this should work.
By having the `.AddContentment()` extension method to all the work,
if a developer uses it, then it will be called twice.
Once by the developer, the other by the Composer,
so I'd need to code more defensively.

By moving the configuration/service-registration back to the Composer,
this will happen only once. Leaving the `.AddContentment()` method to
only handle the user settings.