Added `HasValueFor` IPublishedElement extension method
This is useful (when using ModelsBuilder) for checking if a property
has a value that typically uses an object-type that isn't so easy to
check if it's empty.

e.g. a rich-text editor will use an `IHtmlString` object-type, if the
value is empty, the object-value itself isn't `null`, but contains no
contents. If you wanted to wrap additional markup around the rich-text
contents, you would need to add extra logic to check for empty contents.
But now you can use this... `Model.HasValueFor(x => x.BodyText)`