Data List - Umbraco Content by XPath bug
https://our.umbraco.com/packages/backoffice-extensions/contentment/contentment-feedback/107355-error-when-publishing-document

I can only assume that since Umbraco v8.16.0, the addition of the
`IDisposible` interface to several services has meant that what would
previously be accessible (at a later point in the scope/life-cycle) is
no longer available. So with any lazily-evaluated LINQ results
(particularly for the Content cache) would need to be eagerly-evaluated,
hence the `.ToList()` call here.