Enum data-source no longer binds to specific assembly version
Changed to store the assembly's `Name` as opposed to the `FullName`,
(since this contains the version and culture bindings).

Also since the data-source item's name and value are the same,
I've refactored the WebAPI method to use a `SortedSet`,
as we only need to manage a single value, not Name/FullName combo.

Fixes #197