Added `JsonArrayDataValueEditor`
This ensures that the initial/default value is an empty array,
as opposed to an empty string, which causes issues, e.g.

- https://github.com/leekelleher/umbraco-contentment/issues/109
- https://github.com/leekelleher/umbraco-contentment/issues/242

I have added this data-value-editor to DataList, ListItems, SocialLinks.

I've also done a minor refactor on ContentBlocks's data-value-editor,
as it wasn't quite right.

I've removed the JS edge-case hotfix for ItemPicker and Tags.