DataList - Tracking references
Having previously attempted to get the Tracking References working
with DataList, I'd hit a showstopper, (in that I couldn't detect the
configuration/data-type, so the underlying data-source was an unknown).
However, rather than disregard the feature, I took a halfway compromise,
by taking a cursory look at the DataList value to detect if they are
`Udi`, then tracking the reference.