Adds v8 polyfills for `RawValueProperty` and `PublishedPropertyBase`
In Umbraco v12, an interface update on `IPublishedProperty`,
adding a `GetDeliveryApiValue` method causes a breaking-change with
Contentment's `DetachedPublishedProperty` class.

Reviewing Umbraco core code, I found that the `RawValueProperty` class
is now public (it has been since v9), but I hadn't reviewed the
`DetachedPublishedProperty` when porting to .NET Core, just left it as
it was.  In Umbraco v8, the `RawValueProperty` was marked as internal,
again, I hadn't considered using it.

With Contentment's polyfill code, I am able to backport the
`RawValueProperty` and `PublishedPropertyBase` classes, so to mitigate
the breaking-change introduced by Umbraco v12.

Fixes #322